kisan_app/requirements.txt
alabaster==0.7.10
anaconda-client==1.6.3
anaconda-navigator==1.6.2
anaconda-project==0.6.0
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.22.0
astroid==1.4.9
astropy==1.3.3
awscli==1.11.133
Babel==2.4.0
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
beautifulsoup4==4.6.0
bitarray==0.8.1
bleach==1.5.0
bokeh==0.12.5
boto==2.46.1
botocore==1.6.0
Bottleneck==1.2.1
bz2file==0.98
cffi==1.10.0
chardet==3.0.3
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.7
contextlib2==0.5.5
cryptography==1.8.1
cycler==0.10.0
cymem==1.31.2
Cython==0.25.2
cytoolz==0.8.2
datashape==0.5.4
decorator==4.0.11
dill==0.2.7.1
docutils==0.13.1
entrypoints==0.2.2
et-xmlfile==1.0.1
fastcache==1.0.2
fasttext==0.8.3
Flask==0.12.2
Flask-Cors==3.0.2
ftfy==4.4.3
future==0.16.0
gensim==1.0.1
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.7.1
h5py==2.7.0
HeapDict==1.0.0
html5lib==0.9999999
hyperopt==0.1
idna==2.5
imagesize==0.7.1
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.0.5
langdetect==1.0.7
langid==1.1.6
lazy-object-proxy==1.2.2
lime==0.1.1.23
llvmlite==0.18.0
locket==0.2.0
lxml==3.7.3
Markdown==2.6.8
MarkupSafe==0.23
matplotlib==1.4.3
mistune==0.7.4
mpmath==0.19
msgpack-python==0.4.8
mtranslate==1.6
multipledispatch==0.4.9
murmurhash==0.26.4
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
nltk==3.2.3
nose==1.3.7
notebook==5.0.0
numba==0.33.0
numexpr==2.6.2
numpy==1.13.0
numpydoc==0.6.0
olefile==0.44
opencv-python==3.2.0.7
openpyxl==2.4.7
packaging==16.8
pandas==0.20.2
pandocfilters==1.4.1
partd==0.3.8
pathlib==1.0.1
pathlib2==2.2.1
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
plac==0.9.6
plotly==2.0.14
ply==3.10
preshed==1.0.0
prompt-toolkit==1.0.14
protobuf==3.3.0
psutil==5.2.2
ptyprocess==0.5.1
py==1.4.33
pyasn1==0.3.2
pycosat==0.6.2
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pyemd==0.4.4
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.6.4
pymongo==3.4.0
pyodbc==4.0.16
pyOpenSSL==17.0.0
pyparsing==2.1.4
pytest==3.0.7
python-dateutil==2.6.0
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
regex==2017.7.28
requests==2.14.2
rope-py3k==0.9.4.post1
rsa==3.4.2
s3transfer==0.1.10
scikit-learn==0.18.1
scipy==0.19.0
seaborn==0.7.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.5.2
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
spacy==1.9.0
Sphinx==1.5.6
spyder==3.1.4
SQLAlchemy==1.1.9
sympy==1.0
tables==3.3.0
tblib==1.3.2
tensorflow==1.2.1
termcolor==1.1.0
terminado==0.6
testpath==0.3
tflearn==0.3.2
Theano==0.9.0
thinc==6.5.2
toolz==0.8.2
tornado==4.5.1
tqdm==4.14.0
traitlets==4.3.2
tsne==0.1.7
ujson==1.35
unicodecsv==0.14.1
urllib3==1.21.1
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==2.0.0
wrapt==1.10.10
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.4
xlwt==1.2.0
zict==0.1.2
